+++ TabletopTweaks/Reworks/Aeon.cs
@@ -31,7 +31,7 @@ using TabletopTweaks.NewComponents.AbilitySpecific;
 using TabletopTweaks.NewComponents.Properties;
 using TabletopTweaks.Utilities;
 
-namespace TabletopTweaks.MythicReworks {
+namespace TabletopTweaks.Reworks {
     class Aeon {
         [HarmonyPatch(typeof(BlueprintsCache), "Init")]
         static class BlueprintsCache_Init_Patch {
